Fixes some prefs issues (#11990)
* Fixes some prefs issues

So it turns out SOME morons who've added prefs to the menu in the past never bothered in putting the actual behavior to make the toggles affect the current game before relogging. So I've manually gone through all the prefs to make sure they have the full behavior the verbs did.

* Removes now unused proc/href link

* Last look through