Adds logging when gas-exuding seeds are planted (#12221)
* Adds logging when gas-exuding seeds are planted

Instead of having to check all strange and modified plants for
fingerprints and their seeds for exude_gasses, this'll add a log whenever
a seed which exudes gasses is planted.
Should it add an investigation log to the person planting it? I'm not
sure. I can do that if it should be added, but I'd have to find out how
exactly first.

* Changes plant gas logging to use formatJumpTo()

Doesn't use the old href thing like the other plants do heeeey

* Changes plant gas logging to use add_gamelogs()

Looks neater and has a cool colour!