Fix for grab move logic (#12518)
* Fix for grab move logic

The . = ..() call was removed in #12287 and replaced with a step for the standard case, so I infer it no longer worked to do the step call.

I assume this call in the grab code was meant to also be replaced but was overlooked.

This probably fixes #12516 but I didn't test it yet

* Forgot I changed that